pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.0",
    "pillow>=11.1.0",
    "werkzeug>=3.1.3",
    "psycopg2-binary>=2.9.10",
    "flask-migrate>=4.1.0",
]
